+++ client/src/components/ViewAllAvailabilities.js
@@ -81,7 +81,7 @@ function ViewAllAvailabilities(props) {
 							{availability.username || availability.email}
 						</td> */}
 							<td className="text-center">{new Date(availability.date)
-								.toLocaleDateString("en-GB", { weekday: 'short', year: 'numeric', month: 'long', day: 'numeric' })}
+								.toLocaleDateString("en-GB", { weekday: 'short', year: 'numeric', month: 'short', day: 'numeric' })}
 							</td>
 							<td className="text-center">{new Date(`1970-01-01T${availability.from_time}Z`)
 								.toLocaleTimeString("en-US", { hour12: true, hour: "numeric", minute: "numeric" })}
